Modifications
Segregate the MyApp class into its own file.  Remove Myapp class from xIrc.cpp.  Add the new files into xIrc.pro
Also added the show function in setDefaults function for Defaults object.
